requirements.txt
pdfplumber>=0.10.0
langchain>=0.3.0
langchain-mistralai>=0.2.0
pydantic>=2.0
python-dotenv>=1.0
